refactor(guard): split native policy publisher modules
