+++ src/codex_plugin_scanner/guard/native_policy_snapshot_windows_key.py
@@ -1,0 +1,205 @@
+"""Windows verifier-key persistence using one verified handle."""
+
+from __future__ import annotations
+
+import hmac
+import os
+import stat
+from pathlib import Path
+
+from .native_policy_snapshot_codec import derive_native_policy_verifier_key
+from .native_policy_snapshot_constants import (
+    _VERIFIER_KEY_BYTES,
+    NATIVE_POLICY_VERIFIER_KEY_NAME,
+    NativePolicySnapshotError,
+)
+from .native_policy_snapshot_windows_acl import _windows_verify_private_dacl
+from .native_policy_snapshot_windows_io import (
+    _windows_apply_private_dacl,
+    _windows_close_handle,
+    _windows_open_handle,
+)
+from .native_policy_snapshot_windows_support import (
+    _runtime_state_directory,
+    _windows_owner_sid,
+    _windows_private_descriptor,
+)
+
+
+def _windows_read_key(path: Path, expected: bytes) -> None:
+    import ctypes
+    from ctypes import wintypes
+
+    kernel32, handle, information = _windows_open_handle(path, directory=False)
+    try:
+        owner_sid = _windows_owner_sid()
+        _windows_verify_private_dacl(handle, owner_sid=owner_sid, directory=False)
+        size = (int(information.nFileSizeHigh) << 32) | int(information.nFileSizeLow)
+        if size != _VERIFIER_KEY_BYTES:
+            raise NativePolicySnapshotError("native_policy_verifier_key_invalid")
+        read_file = kernel32.ReadFile
+        read_file.argtypes = [
+            wintypes.HANDLE,
+            ctypes.c_void_p,
+            wintypes.DWORD,
+            ctypes.POINTER(wintypes.DWORD),
+            ctypes.c_void_p,
+        ]
+        read_file.restype = wintypes.BOOL
+        buffer = (ctypes.c_ubyte * (_VERIFIER_KEY_BYTES + 1))()
+        count = wintypes.DWORD()
+        if (
+            not read_file(
+                handle,
+                buffer,
+                _VERIFIER_KEY_BYTES + 1,
+                ctypes.byref(count),
+                None,
+            )
+            or count.value != _VERIFIER_KEY_BYTES
+        ):
+            raise NativePolicySnapshotError("native_policy_verifier_key_read_failed")
+        if not hmac.compare_digest(bytes(buffer[: count.value]), expected):
+            raise NativePolicySnapshotError("native_policy_verifier_key_mismatch")
+    finally:
+        _windows_close_handle(kernel32, handle)
+
+
+def _windows_provision_verifier_key(path: Path, derived: bytes) -> Path:
+    import ctypes
+    from ctypes import wintypes
+
+    try:
+        _windows_read_key(path, derived)
+        return path
+    except FileNotFoundError:
+        pass
+    owner_sid = _windows_owner_sid()
+    try:
+        with _windows_private_descriptor(False) as (_advapi32, descriptor, dacl, _descriptor_owner_sid):
+            kernel32, handle, _information = _windows_open_handle(
+                path,
+                directory=False,
+                create_new=True,
+                descriptor=descriptor,
+            )
+            try:
+                write_file = kernel32.WriteFile
+                write_file.argtypes = [
+                    wintypes.HANDLE,
+                    ctypes.c_void_p,
+                    wintypes.DWORD,
+                    ctypes.POINTER(wintypes.DWORD),
+                    ctypes.c_void_p,
+                ]
+                write_file.restype = wintypes.BOOL
+                buffer = (ctypes.c_ubyte * len(derived)).from_buffer_copy(derived)
+                count = wintypes.DWORD()
+                if not write_file(
+                    handle,
+                    buffer,
+                    len(derived),
+                    ctypes.byref(count),
+                    None,
+                ) or count.value != len(derived):
+                    raise NativePolicySnapshotError("native_policy_verifier_key_write_failed")
+                flush = kernel32.FlushFileBuffers
+                flush.argtypes = [wintypes.HANDLE]
+                flush.restype = wintypes.BOOL
+                if not flush(handle):
+                    raise NativePolicySnapshotError("native_policy_verifier_key_sync_failed")
+                _windows_apply_private_dacl(kernel32, handle, descriptor, dacl, False)
+                _windows_verify_private_dacl(handle, owner_sid=owner_sid, directory=False)
+            finally:
+                _windows_close_handle(kernel32, handle)
+    except FileExistsError:
+        _windows_read_key(path, derived)
+        return path
+    except NativePolicySnapshotError:
+        raise
+    except (OSError, RuntimeError, TypeError, ValueError) as error:
+        raise NativePolicySnapshotError("native_policy_verifier_key_write_failed") from error
+    return path
+
+
+def provision_native_policy_verifier_key(
+    guard_home: Path,
+    policy_integrity_key: bytes,
+) -> Path:
+    """Provision the owner-private derived key consumed by the Rust resident.
+
+    Creation uses ``O_EXCL`` and never replaces an existing key. A changed
+    policy master is therefore a safe failure requiring explicit repair rather
+    than an implicit verifier reset that could invalidate the resident floor.
+    """
+
+    derived = derive_native_policy_verifier_key(policy_integrity_key)
+    state_dir = _runtime_state_directory(guard_home)
+    path = state_dir / NATIVE_POLICY_VERIFIER_KEY_NAME
+    if os.name == "nt":
+        return _windows_provision_verifier_key(path, derived)
+    try:
+        metadata = path.lstat()
+    except FileNotFoundError:
+        metadata = None
+    except OSError as error:
+        raise NativePolicySnapshotError("native_policy_verifier_key_stat_failed") from error
+
+    if metadata is not None:
+        if stat.S_ISLNK(metadata.st_mode) or not stat.S_ISREG(metadata.st_mode):
+            raise NativePolicySnapshotError("native_policy_verifier_key_invalid")
+        if metadata.st_size != _VERIFIER_KEY_BYTES:
+            raise NativePolicySnapshotError("native_policy_verifier_key_invalid")
+        if os.name != "nt" and (metadata.st_uid != os.geteuid() or stat.S_IMODE(metadata.st_mode) & 0o077):
+            raise NativePolicySnapshotError("native_policy_verifier_key_not_private")
+        flags = os.O_RDONLY | getattr(os, "O_CLOEXEC", 0) | getattr(os, "O_NOFOLLOW", 0)
+        try:
+            descriptor = os.open(path, flags)
+        except OSError as error:
+            raise NativePolicySnapshotError("native_policy_verifier_key_read_failed") from error
+        try:
+            opened = os.fstat(descriptor)
+            existing = os.read(descriptor, _VERIFIER_KEY_BYTES + 1)
+        finally:
+            os.close(descriptor)
+        if (
+            not stat.S_ISREG(opened.st_mode)
+            or opened.st_size != _VERIFIER_KEY_BYTES
+            or len(existing) != _VERIFIER_KEY_BYTES
+            or not hmac.compare_digest(existing, derived)
+        ):
+            raise NativePolicySnapshotError("native_policy_verifier_key_mismatch")
+        return path
+
+    flags = os.O_WRONLY | os.O_CREAT | os.O_EXCL | getattr(os, "O_CLOEXEC", 0) | getattr(os, "O_NOFOLLOW", 0)
+    try:
+        descriptor = os.open(path, flags, 0o600)
+    except FileExistsError:
+        # Another process won the race. Re-enter the validation path without
+        # replacing or weakening its key.
+        return provision_native_policy_verifier_key(guard_home, policy_integrity_key)
+    except OSError as error:
+        raise NativePolicySnapshotError("native_policy_verifier_key_write_failed") from error
+    try:
+        written = 0
+        while written < len(derived):
+            written += os.write(descriptor, derived[written:])
+        os.fsync(descriptor)
+    except OSError as error:
+        raise NativePolicySnapshotError("native_policy_verifier_key_write_failed") from error
+    finally:
+        os.close(descriptor)
+    if os.name != "nt":
+        try:
+            path.chmod(0o600)
+            directory_descriptor = os.open(
+                state_dir,
+                os.O_RDONLY | getattr(os, "O_DIRECTORY", 0) | getattr(os, "O_CLOEXEC", 0),
+            )
+            try:
+                os.fsync(directory_descriptor)
+            finally:
+                os.close(directory_descriptor)
+        except OSError as error:
+            raise NativePolicySnapshotError("native_policy_verifier_key_sync_failed") from error
+    return path
